refactor(emitter): split class ES5 comment helpers
